+++ pytorch_lightning/trainer/connectors/checkpoint_connector.py
@@ -226,7 +226,6 @@ class CheckpointConnector:
         """Restores all callbacks from the pre-loaded checkpoint."""
         if not self._loaded_checkpoint:
             return
-        self.trainer._check_ckpt_callbacks_presence_on_load(self._loaded_checkpoint)
         self.trainer._call_callbacks_on_load_checkpoint(self._loaded_checkpoint)
         self.trainer._call_callbacks_load_state_dict(self._loaded_checkpoint)
 
@@ -321,10 +320,8 @@ class CheckpointConnector:
                 'epoch':                     training epoch
                 'global_step':               training global step
                 'pytorch-lightning_version': The version of PyTorch Lightning that produced this checkpoint
-                'callbacks_state_dict':                                  # if not weights_only
-                    {callback.state_key: callback.state_dict}
-                'callbacks_deprecated_hook_states':                      # if not weights_only
-                    {callback.state_key: callback.deprecated_on_save_checkpoint}
+                'callbacks':                                  # if not weights_only
+                    {callback.state_key: callback.on_save_checkpoint or callback.state_dict}
                 'optimizer_states':          "PT optim's state_dict"[]   # if not weights_only
                 'lr_schedulers':             "PT sched's state_dict"[]   # if not weights_only
                 'state_dict':                Model's state_dict (e.g. network weights)
@@ -349,7 +346,7 @@ class CheckpointConnector:
 
         if not weights_only:
             # dump callbacks
-            checkpoint["callbacks_state_dict"] = self.trainer._call_callbacks_state_dict()
+            checkpoint["callbacks"] = self.trainer._call_callbacks_state_dict()
 
             optimizer_states = []
             for i, optimizer in enumerate(self.trainer.optimizers):
@@ -392,7 +389,7 @@ class CheckpointConnector:
 
         # on_save_checkpoint hooks
         if not weights_only:
-            checkpoint["callbacks_deprecated_hook_states"] = self.trainer._call_callbacks_on_save_checkpoint(checkpoint)
+            self.trainer._call_callbacks_on_save_checkpoint(checkpoint)
         model.on_save_checkpoint(checkpoint)
         if datamodule is not None:
             datamodule.on_save_checkpoint(checkpoint)